changing bigint fields to numeric(78)

+++ lndr-backend/db/create_tables.sql
@@ -1,20 +1,20 @@
 CREATE TABLE pending_credits (
     hash        CHAR(64) PRIMARY KEY,
     submitter   CHAR(40),
-    nonce       BIGINT,
+    nonce       NUMERIC(78),
     creditor    CHAR(40),
     debtor      CHAR(40),
-    amount      BIGINT,
+    amount      NUMERIC(78),
     memo        CHAR(32),
     signature   CHAR(130)
 );
 
 CREATE TABLE verified_credits (
     hash               CHAR(64) PRIMARY KEY,
-    nonce              BIGINT,
+    nonce              NUMERIC(78),
     creditor           CHAR(40),
     debtor             CHAR(40),
-    amount             BIGINT,
+    amount             NUMERIC(78),
     memo               CHAR(32),
     creditor_signature CHAR(130),
     debtor_signature   CHAR(130)
